==> 2. Preprocessing Hasil Scraping Linkedin.ipynb <==
# Preprocessing Hasil Scraping LinkedIn

import pandas as pd
import numpy as np
import json

df = pd.read_csv("Scraping Profesi Data (Linkedin).csv")

df.head()

len(df)

# Mengubah df["icon_keterangan"] agar menjadi dict, bukan str
for i in range(0, len(df)):
    s = df["icon_keterangan"][i]
    json_acceptable_string = s.replace("'", "\"")
    d = json.loads(json_acceptable_string)
    df["icon_keterangan"][i] = d

# Gabungkan dengan data frame awal
df_gabung = pd.concat([df.drop("icon_keterangan", axis = 1), pd.DataFrame(df["icon_keterangan"].tolist())], axis = 1)
df_gabung.head()

# Preprocessing 1

### Untuk Kolom "briefcase-icon" & "company-icon"

# GABUNG FOR LOOPING UNTUK "briefcase-icon" & "company-icon"

list_dict_gab = []

# Data frame awal
df = pd.DataFrame(columns = ["gaji", "jenis_job", "tingkat_job", "ukuran_company", "industri"])
    
for i in range(0, len(df_gabung)):
    dict_gab = {}
    
    # Untuk kolom "briefcase-icon"
    if pd.isna(df_gabung["briefcase-icon"][i]) == True:
        brief_split = {"jenis_job": np.nan}
    else:
        split_1 = df_gabung["briefcase-icon"][i].split(" · ")
        if len(split_1) == 1:
            brief_split = {"jenis_job": split_1[0]}
        elif len(split_1) == 2:
            brief_split = {"jenis_job": split_1[0], "tingkat_job": split_1[1]}
        elif len(split_1) == 3:
            brief_split = {"gaji": split_1[0], "jenis_job": split_1[1], "tingkat_job": split_1[2]}
            
    dict_gab.update(brief_split) 
    
    # Untuk kolom "company-icon"
    if pd.isna(df_gabung["company-icon"][i]) == True:
        company_split = {"ukuran_company": np.nan}
    else:
        split_2 = df_gabung["company-icon"][i].split(" · ")
        if len(split_2) == 1:
            company_split = {"ukuran_company": split_2[0]}
        elif len(split_2) == 2:
            company_split = {"ukuran_company": split_2[0], "industri": split_2[1]}
    
    dict_gab.update(company_split) 
    
    list_dict_gab.append(dict_gab) # Nanti yang digunakan jadi "list_dict_gab"-nya

list_dict_gab

#df_gabung2 = pd.concat([df_gabung.drop(["briefcase-icon", "company-icon"], axis = 1), pd.DataFrame(list_dict_gab)], axis = 1)
df_gabung2 = pd.concat([df_gabung, pd.DataFrame(list_dict_gab)], axis = 1)
df_gabung2 = df_gabung2.drop(["briefcase-icon", "company-icon", "radar-screen-icon", "people-icon"], axis = 1)
df_gabung2.head()

#df_gabung2.to_csv("Hasil Preprocessing 1 (Kolom OKE).csv", index = False)

# Preprocessing 2

### Ganti Job Name dan Lokasi

data = pd.read_csv("Hasil Preprocessing 1 (Kolom OKE).csv")
data.head(3)

data.info()

# Cek data yang duplikat
data[data.duplicated(keep = False)] 

# Data Analyst/ Data Scientist -> Ini dihapus saja agar tidak bingung mau dimasukkan ke "data analyst" atau "data scientist"

data.drop_duplicates(keep = False, inplace = True, ignore_index = True)
data.info() # Ada 4 yang terhapus

data[data.duplicated(keep = False)] # Sudah tidak ada yang duplikat

data.index # Index-nya sudah urut

# PROSESNYA DIGABUNG MENJADI SATU

def ganti_nama_dan_lokasi(data, list_kolom):
    data2 = data.copy()
    
    # Looping untuk lower case isi dari masing-masing kolom
    for kolom in list_kolom:
        data2[kolom] = data2[kolom].str.lower()

    # Looping untuk ganti isi di kolom "job_name" dan "lokasi"
    for i in range(0, len(data2)):
        # Untuk kolom "job_name"
        # ["data scientist", "data analyst", "data engineer", "database manager", "data architect", "database administrator"]
        if "data scientist" in data2["job_name"][i]:
            data2["job_name"][i] = "data scientist"
        elif "data analyst" in data2["job_name"][i]:
            data2["job_name"][i] = "data analyst"
        elif "data engineer" in data2["job_name"][i]:
            data2["job_name"][i] = "data engineer"
        elif "database manager" in data2["job_name"][i]:
            data2["job_name"][i] = "database manager"
        elif "data architect" in data2["job_name"][i]:
            data2["job_name"][i] = "data architect"
        elif "database administrator" in data2["job_name"][i]:
            data2["job_name"][i] = "database administrator"

        # Untuk kolom "lokasi"
        if "jakarta" in data2["lokasi"][i]:
            data2["lokasi"][i] = "dki jakarta"
        elif "banten" in data2["lokasi"][i]:
            data2["lokasi"][i] = "banten"
        elif "yogyakarta" in data2["lokasi"][i]:
            data2["lokasi"][i] = "diy"
        elif "kalimantan barat" in data2["lokasi"][i]:
            data2["lokasi"][i] = "kalimantan barat"
        elif "sulawesi selatan" in data2["lokasi"][i]:
            data2["lokasi"][i] = "sulawesi selatan"
        elif "jawa barat" in data2["lokasi"][i]:
            data2["lokasi"][i] = "jawa barat"
        elif "jawa tengah" in data2["lokasi"][i]:
            data2["lokasi"][i] = "jawa tengah"
        elif "jawa timur" in data2["lokasi"][i]:
            data2["lokasi"][i] = "jawa timur"
        elif "nusa tenggara barat" in data2["lokasi"][i]:
            data2["lokasi"][i] = "ntb"
        elif "bali" in data2["lokasi"][i]:
            data2["lokasi"][i] = "bali"
        elif "nanggroe aceh darussalam" in data2["lokasi"][i]:
            data2["lokasi"][i] = "nad"
    
    return data2

list_kolom = ["job_name", "lokasi", "jenis_job", "tingkat_job", "ukuran_company", "industri", "gaji"]

data_oke = ganti_nama_dan_lokasi(data, list_kolom)
data_oke.head()

data_oke[data_oke["lokasi"] == "indonesia"] # Akan dihapus karena base-nya di Singapura dan di Indonesia juga sistemnya remote

data_oke.drop(250, axis = 0, inplace = True)
data_oke = data_oke.reset_index(drop = True)
len(data_oke) # Sudah berkurang satu

data_oke.index

#data_oke.to_csv("Hasil Preprocessing 2 (Ganti Job Name dan Lokasi).csv", index = False)

# Preprocessing 3

### Memperbaiki Kolom Gaji & Membuat Kolom Median Gaji

data_oke = pd.read_csv("Hasil Preprocessing 2 (Ganti Job Name dan Lokasi).csv")
data_oke.tail()

data_oke["gaji"].unique()

data_oke2 = data_oke.copy()

# GAJI PER BULAN
for i in range(0, len(data_oke2)):
    if pd.isna(data_oke2["gaji"][i]) == True:
        pass
    else:
        if "month" in data_oke2["gaji"][i]:
            data_oke2["gaji"][i] = data_oke2["gaji"][i].replace("/month", "").replace("rp", "").replace(".", "")
            data_oke2["gaji"][i] = data_oke2["gaji"][i].split(" - ")
            
            if type(pd.isna(data_oke2["gaji"][i])) == type(True):
                pass
            elif len(data_oke2["gaji"][i]) == 2: # Jika list-nya berisi 2 nilai
                list_angka = [int(angka) for angka in data_oke2["gaji"][i]]
                data_oke2["gaji"][i] = list_angka
            else:
                pass
            
        elif "yr" in data_oke2["gaji"][i]:
            data_oke2["gaji"][i] = data_oke2["gaji"][i].replace("/yr", "").replace("rp", "").replace(".", "")
            data_oke2["gaji"][i] = data_oke2["gaji"][i].split(" - ")
            #data_oke2["gaji"][i] = [tahun/12 for tahun in data_oke2["gaji"][i]]
            
            if type(pd.isna(data_oke2["gaji"][i])) == type(True):
                pass
            elif len(data_oke2["gaji"][i]) == 2: # Jika list-nya berisi 2 nilai
                list_angka = [round(int(angka)/12) for angka in data_oke2["gaji"][i]]
                data_oke2["gaji"][i] = list_angka
            else:
                pass

for i in range(0, len(data_oke2)):
    if type(data_oke2["gaji"][i]) == list:
        print(i)
        print(data_oke2["gaji"][i])
    
# Yang belum beres -> index 25, 42, 150, 196, 217

data_oke.iloc[[25, 42, 150, 196, 217]] 

# Diasumsikan yang hanya tertulis 5, 20, 10, dst itu satuannya juta

data_oke2["gaji"][25]

data_oke2["gaji"][150]

# Menangani yang bentuk range (index 25, 42, 196)
data_oke2["gaji"][25] = [5000000 if item == 5 else 20000000 for item in data_oke2["gaji"][25]]
data_oke2["gaji"][42] = [15000000 if item == 15 else 0 for item in data_oke2["gaji"][42]]
data_oke2["gaji"][196] = [10000000 if item == 10 else 20000000 for item in data_oke2["gaji"][196]]

# Menangani yang isi list-nya cuma 1 (index 150, 217)
data_oke2["gaji"][150] = [int(data_oke2["gaji"][150][0].replace("sgd300", "3146244"))] # 300 SGD -> Rp 3.146.244,-
data_oke2["gaji"][217] = [round(int(data_oke2["gaji"][217][0].replace("1", "1000000")) / 12)]

data_oke2.iloc[[25, 42, 150, 196, 217]]

# Menambah kolom baru untuk median dari gaji
import statistics

median_gaji = []
for i in range(0, len(data_oke2)):
    if type(data_oke2["gaji"][i]) != list:
        median_gaji.append(np.nan)
    else:
        median_gaji.append(round(statistics.median(data_oke2["gaji"][i])))
        
data_oke2["median_gaji"] = median_gaji

data_oke2.tail()

#data_oke2.to_csv("Hasil Preprocessing 3 (Benerin Kolom Gaji & Bikin Kolom Median Gaji).csv", index = False)

# Preprocessing 4

### Menyederhanakan Kolom Industri

data_oke2 = pd.read_csv("Hasil Preprocessing 3 (Benerin Kolom Gaji & Bikin Kolom Median Gaji).csv")
data_oke2.tail()

data_oke2["industri"].nunique()

# Misal:
# Keuangan -> 'Perbankan', 'Jasa Keuangan'
# Pendidikan -> 'Belajar Online', 'Manajemen Pendidikan'
# Teknologi Informasi dan Komunikasi -> 'Layanan dan Teknologi Informasi', 'Internet', 'Telekomunikasi', 'Perangkat Lunak Komputer'
# Media -> 'Media Online'
# Pariwisata -> 'Wisata, Travel & Pariwisata', 'Pariwisata & Perhotelan'
# Kesehatan -> 'Perawatan Kesehatan Jiwa', 'Kesehatan dan Kebugaran', 'Olahraga'
# Agrikultur -> 'Kertas & Produk Hutan', 'Pertanian', 'Tembakau'
# Human Resources -> 'Penetapan Staf dan Perekrutan', 'Sumber Daya Manusia'
# Makanan & Minuman -> 'Makanan & Minuman', 'Wine dan Alkohol'
# Ekonomi -> 'Logistik dan Rantai Suplai', 'Modal Usaha & Ekuitas Pribadi', 'Ritel'

# Ini yang sendiri saja:
# 'Otomotif', 'Kosmetik', 'Pakaian & Mode', 'Minyak & Energi', 'Perabot', 
# 'Konsultan Manajemen' (Industri Konsultasi), 
# 'Pemasaran dan Periklanan' (Industri Pemasaran & Periklanan)

industri_dict = {"perbankan": "keuangan", "jasa keuangan": "keuangan", 
                 "belajar online": "pendidikan", "manajemen pendidikan": "pendidikan", 
                 "layanan dan teknologi informasi": "teknologi informasi dan komunikasi", "internet": "teknologi informasi dan komunikasi", "telekomunikasi": "teknologi informasi dan komunikasi", 
                 "perangkat lunak komputer": "teknologi informasi dan komunikasi", "media online": "media", 
                 "wisata, travel & pariwisata": "pariwisata", "pariwisata & perhotelan": "pariwisata", 
                 "perawatan kesehatan jiwa": "kesehatan", "kesehatan dan kebugaran": "kesehatan", "olahraga": "kesehatan", 
                 "kertas & produk hutan": "agrikultur", "tembakau": "agrikultur", "pertanian": "agrikultur",
                 "penetapan staf dan perekrutan": "human resources", "sumber daya manusia": "human resources", 
                 "wine dan alkohol": "makanan & minuman", 
                 "logistik dan rantai suplai": "ekonomi", "modal usaha & ekuitas pribadi": "ekonomi", "ritel": "ekonomi", 
                 "konsultan manajemen": "konsultasi", "pemasaran dan periklanan": "pemasaran & periklanan", 
                 "otomotif": "permesinan", "minyak & energi": "pertambangan"}

data_oke3 = data_oke2.copy()
data_oke3.head()

for i in range(0, len(data_oke3)):
    if pd.isna(data_oke3["industri"][i]) == True:
        pass
    else:
        for awal, pengganti in industri_dict.items():
            if awal in data_oke3["industri"][i]:
                data_oke3["industri"][i] = data_oke3["industri"][i].replace(awal, pengganti)

data_oke3.head()

data_oke3["industri"].nunique()

#data_oke3.to_csv("Hasil Preprocessing 4 (Menyederhanakan Kolom Industri).csv", index = False)

# Preprocessing 5

### Menyederhanakan Kolom Ukuran Company

data_oke4 = pd.read_csv("Hasil Preprocessing 4 (Menyederhanakan Kolom Industri).csv")
data_oke4.tail()

data_oke4["ukuran_company"].unique()

# "karyawan" diubah menjadi "pekerja"
data_oke4["ukuran_company"] = data_oke4["ukuran_company"].str.replace("karyawan", "pekerja")

data_oke4["ukuran_company"].unique()

# Disesuaikan dengan data dari Jobstreet juga, sehingga ada yang diubah

for i in range(0, len(data_oke4)):
    if data_oke4["ukuran_company"][i] == "1-10 pekerja":
        data_oke4.iloc[i] = data_oke4.iloc[i].replace("1-10 pekerja", "1-50 pekerja")
    elif data_oke4["ukuran_company"][i] == "11-50 pekerja":
        data_oke4.iloc[i] = data_oke4.iloc[i].replace("11-50 pekerja", "1-50 pekerja")
    
    elif data_oke4["ukuran_company"][i] == "5.001-10.000 pekerja":
        data_oke4.iloc[i] = data_oke4.iloc[i].replace("5.001-10.000 pekerja", ">5.000 pekerja")
    elif data_oke4["ukuran_company"][i] == "10.001+ pekerja":
        data_oke4.iloc[i] = data_oke4.iloc[i].replace("10.001+ pekerja", ">5.000 pekerja")

data_oke4["ukuran_company"].unique()

#data_oke4.to_csv("Hasil Preprocessing 5 (Menyederhanakan Kolom Ukuran Company).csv", index = False)

######## INSYAALLOH FIX PAKAI INI (Hasil Preprocessing 5)

len(data_oke4)
