==> 3. Gabungan Data Linkedin dan Jobstreet (Hasil Preprocessing).ipynb <==
# Menggabungkan Data Linkedin dan Jobstreet (Hasil Preprocessing)

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data_linkedin = pd.read_csv("Hasil Preprocessing 5 (Menyederhanakan Kolom Ukuran Company).csv")
data_jobstreet = pd.read_csv("Hasil Preprocessing 7 (Memperbaiki Kolom Lama Pengalaman) Data Jobstreet.csv")

data_linkedin["sumber"] = "linkedin"
data_jobstreet["sumber"] = "jobstreet"

data_linkedin.head(2)

len(data_linkedin)

data_jobstreet.head(2)

len(data_jobstreet)

df_gabung = data_jobstreet.append(data_linkedin)
df_gabung = df_gabung.reset_index(drop = True)
df_gabung.tail()

len(df_gabung)

df_gabung["job_name"].value_counts()

# Jadikan lower case untuk kolom "company"
df_gabung["company"] = df_gabung["company"].str.lower()

df_gabung.tail()

# Disamakan perusahaannya (ada yang penulisannya beda) 
# -> Nanti akan berguna untuk mengisi missing value di kolom "ukuran_company" dan untuk mengecek duplikat

dict_company = {"pt kamoro maxima integra": "pt kamoro maxima integra", 
                "impact credit solutions": "impact credit solutions", 
                "pt. sharing vision indonesia": "pt sharing vision indonesia", 
                "pt agriaku digital indonesia": "agriaku", 
                "pt takjub teknologi indonesia (ajaib)": "ajaib", 
                "kompas gramedia": "kompas gramedia - kg media", 
                "detik.com": "detikcom", 
                "j-express": "j-express (pt jaya ekspress transindo)", 
                "pt lawencon": "pt lawencon internasional", 
                "pt. moladin digital indonesia": "moladin", 
                "pt tani hub indonesia (tanihub, tanifund)": "pt tani hub indonesia", 
                "boga group": "boga group", 
                "warung pintar group": "warung pintar", 
                "gramedia e-perpust": "kompas gramedia - kg media", 
                "gramedia writing project": "kompas gramedia - kg media", 
                "liv.it": "livit international", 
                "cicil": "pt cicil solusi mitra teknologi", 
                "pt. metrodata electronics, tbk": "pt metrodata electronics tbk",
                "pt. metrodata electronics tbk.": "pt metrodata electronics tbk",
                "pt metrodata electronics tbk": "pt metrodata electronics tbk", 
                "ritase": "ritase.com", 
                "zalora south east asia pte ltd": "zalora group", 
                "99.co indonesia": "99.co", 
                "amartha": "pt. amartha mikro fintek", 
                "ntt indonesia digital business solutions": "ntt ltd.", 
                "higo": "pt higo fitur indonesia", 
                "bobobox indonesia": "pt bobobox mitra indonesia", 
                "jobs mnc": "mnc group",
                "mnc portal indonesia": "mnc group", 
                "bitlabs": "bitlabs academy",
                "mitramas infosys global": "pt mitramas infosys global", 
                "dhana jasa konsultan": "pt. dhana jasa konsultan", 
                "modal rakyat": "modal rakyat indonesia", 
                "generasi teknologi buana": "pt generasi teknologi buana"} 

for i in range(0, len(df_gabung)):
    if pd.isna(df_gabung["company"][i]) == True:
        pass
    else:
        for awal, pengganti in dict_company.items():
            if awal in df_gabung["company"][i]:
                df_gabung["company"][i] = df_gabung["company"][i].replace(awal, pengganti)

### Baca Data Gabungan

df = df_gabung.copy()
df.head(3)

# Cek data yang duplikat
duplikat = df[df[["job_name", "company", "lokasi", "tingkat_job", "lama_pengalaman", "jenis_job", "ukuran_company", "industri", "gaji"]].duplicated(keep = False)] 
# keep = False -> agar bisa kelihatan semua datanya (termasuk data asli dan duplikatnya)

#duplikat.to_csv("Duplikat dari Data Gabungan.csv", index = False)

df.info()

df.drop_duplicates(keep = "first", inplace = True, ignore_index = True) 
# keep = "first" -> Pertahankan data yang pertama saja, sisanya dianggap duplikat dan dihapus 

df.info()

# Save to csv
df.to_csv("Data Gabungan dari Linkedin dan Jobstreet (3 Profesi).csv", index = False)
