==> 4. EDA dari Data Gabungan.ipynb <==
# EDA Data Gabungan

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data = pd.read_csv("Data Gabungan dari Linkedin dan Jobstreet (3 Profesi).csv")

data.head(2)

data.info()

len(data)

data.describe()

def summarize_features(df):
    # first column will be data types of each feature
    summary = pd.DataFrame(df.dtypes, columns = ['dtypes'])
    summary = summary.reset_index()
    # how many missing values in each feature
    summary['Missing'] = df.isnull().sum().values
    # how many unique values in each feature (cardinality indicator)
    summary['Uniques'] = df.nunique().values

    return summary

summarize_features(data)

### 1. Frekuensi Nama Profesi Data

data["job_name"].unique()

size_job_name = data.groupby(["job_name"]).size().sort_values(ascending = False)
size_job_name

x = size_job_name.index
y = size_job_name.values

plt.bar(x, y)
plt.title("Frekuensi Nama Profesi Data")
plt.xlabel("Nama Profesi Data")
plt.ylabel("Frekuensi")
plt.xticks(x)

plt.savefig("gambar/1. Frekuensi Nama Profesi Data.png")

plt.show()

### 2. Frekuensi Lokasi Perusahaan

data["lokasi"].unique()

size_lokasi = data.groupby(["lokasi"]).size().sort_values(ascending = False)
size_lokasi

size_lokasi.head(7).index

plt.figure(figsize = (9, 5))

x = size_lokasi.head(7).index
y = size_lokasi.head(7).values

plt.bar(x, y)
plt.title("Frekuensi Lokasi Perusahaan")
plt.xlabel("Lokasi")
plt.ylabel("Frekuensi")
plt.xticks(x)

plt.savefig("gambar/2. Frekuensi Lokasi Perusahaan.png")

plt.show()

### 3. Frekuensi Tingkat Pekerjaan

data["tingkat_job"].unique()

size_tingkat_job = data.groupby(["tingkat_job"]).size().sort_values(ascending = False)
size_tingkat_job

plt.figure(figsize = (8, 9))

x = size_tingkat_job.index
y = size_tingkat_job.values

plt.bar(x, y)
plt.title("Frekuensi Tingkat Pekerjaan")
plt.xlabel("Tingkat Pekerjaan")
plt.ylabel("Frekuensi")
plt.xticks(x, rotation = 20)

plt.savefig("gambar/3. Frekuensi Tingkat Pekerjaan.png")

plt.show()

### 4. Rata-Rata Lama Pengalaman

data["lama_pengalaman"][~np.isnan(data["lama_pengalaman"])]

filtered_lama_pengalaman = data["lama_pengalaman"][~np.isnan(data["lama_pengalaman"])]

plt.boxplot(filtered_lama_pengalaman, meanline = True, showmeans = True)
plt.title("Boxplot Lama Pengalaman")
plt.ylabel("Lama Pengalaman (Tahun)")

plt.savefig("gambar/4. Boxplot Lama Pengalaman.png")

plt.show()

### 5. Frekuensi Jenis Pekerjaan

data["jenis_job"].unique()

size_jenis_job = data.groupby(["jenis_job"]).size().sort_values(ascending = False)
size_jenis_job

x = size_jenis_job.index
y = size_jenis_job.values

plt.bar(x, y)
plt.title("Frekuensi Jenis Pekerjaan")
plt.xlabel("Jenis Pekerjaan")
plt.ylabel("Frekuensi")
plt.xticks(x)

plt.savefig("gambar/5. Frekuensi Jenis Pekerjaan.png")

plt.show()

### 6. Frekuensi Ukuran Perusahaan

data["ukuran_company"].unique()

size_ukuran_company = data.groupby(["ukuran_company"]).size().sort_values(ascending = False)
size_ukuran_company

plt.figure(figsize = (8, 9))

x = size_ukuran_company.index
y = size_ukuran_company.values

plt.bar(x, y)
plt.title("Frekuensi Ukuran Perusahaan")
plt.xlabel("Ukuran Perusahaan")
plt.ylabel("Frekuensi")
plt.xticks(x, rotation = 20)

plt.savefig("gambar/6. Frekuensi Ukuran Perusahaan.png")

plt.show()

### 7. Frekuensi Industri

data["industri"].unique()

size_industri = data.groupby(["industri"]).size().sort_values(ascending = False)
size_industri

plt.figure(figsize = (12, 15))

x = size_industri.head(6).index
y = size_industri.head(6).values

plt.bar(x, y)
plt.title("Frekuensi Industri")
plt.xlabel("Industri")
plt.ylabel("Frekuensi")
plt.xticks(x, rotation = 20)

plt.savefig("gambar/7. Frekuensi Industri.png")

plt.show()

### 8. Rata-Rata Gaji

data["median_gaji"][~np.isnan(data["median_gaji"])]

filtered_median_gaji = data["median_gaji"][~np.isnan(data["median_gaji"])]

plt.boxplot(filtered_median_gaji, meanline = True, showmeans = True)
plt.title("Boxplot Gaji")
plt.ylabel("Gaji (Rp per Bulan)")

plt.savefig("gambar/8. Boxplot Gaji.png")

plt.show()
